requirements.txt
pyserial
cryptography>=41.0.0
pycryptodome
pystray
Pillow